Refactor to a more domain centered structure. Layers by package: * Adapters ** Primary (driving) adapters, e.g. REST ** Secondary (driven) adapters, e.g. persistence * Application - Takes care of using persistent storage, provides use-case-based methods * Domain - Business logic without any notion of underlying framework etc. See "https://web.archive.org/web/20180822100852/http://alistair.cockburn.us/Hexagonal+architecture"
