+++ src/main/java/com/github/manedev79/timesheet/application/WorkingDayDto.java
@@ -28,12 +28,12 @@ public class WorkingDayDto {
 
     private List<BreakDto> breaks;
 
-    WorkingDay toEntity() {
+    public WorkingDay toEntity() {
         return new WorkingDay(id, day, start, end, description,
                 breaks.stream().map(dto -> dto.fromDto(id)).collect(toList()));
     }
 
-    static WorkingDayDto toDto(final WorkingDay workingDay) {
+    public static WorkingDayDto toDto(final WorkingDay workingDay) {
 
         final List<BreakDto> breakDtos = workingDay.getBreaks().stream()
                 .map(BreakDto::toDto)
